Moved classes to appropriate places
