+++ dsl/src/test/scala/ru/itclover/tsp/dsl/ASTTest.scala
@@ -4,7 +4,7 @@ import org.scalatest.prop.PropertyChecks
 import org.scalatest.{FlatSpec, Matchers}
 import ru.itclover.tsp.core.Intervals.TimeInterval
 import ru.itclover.tsp.core.Window
-import ru.itclover.tsp.utils.UtilityTypes.ParseException
+import UtilityTypes.ParseException
 
 import scala.reflect.ClassTag
