+++ http/src/main/scala/ru/itclover/tsp/http/utils/Exceptions.scala
@@ -1,11 +1,9 @@
-package ru.itclover.tsp.utils
+package ru.itclover.tsp.http.utils
 
 import java.io.{PrintWriter, StringWriter}
 
 object Exceptions {
   case class InvalidRequest(msg: String) extends RuntimeException(msg)
-
-  case class SourceException(msg: String) extends RuntimeException(msg)
 
   def getStackTrace(t: Throwable): String = {
     val sw = new StringWriter()
